--- js/.gitignore
@@ -1,6 +1,0 @@
-/1.1.bundle.js
-/2.2.bundle.js
-/3.3.bundle.js
-/admin.bundle.js
-/main.bundle.js
-/tests.bundle.js
