+++ js/components/interface/plot/configuration/plotConfiguration.js
@@ -1,57 +1,70 @@
-var defaultTrace = {
-    mode: "lines",
-    line: {
-	dash: 'solid',
-	width: '2'
+export const defaultConfig = () => ({ 
+  scrollZoom: false, 
+  displaylogo: false, 
+  modeBarButtonsToRemove: [
+    'resetScale2d', 
+    'hoverClosestCartesian', 
+    'hoverCompareCartesian', 
+    'toggleSpikelines'] 
+})
+  
+  
+export const defaultLine = () => ({
+  dash: 'solid',
+  width: '2'
+})
+  
+export const defaultTrace = () => ({
+  mode: "lines",
+  line: defaultLine(),
+  hoverinfo: 'all',
+  type: 'scatter'
+})
+  
+const defaultFont = () => 'Helvetica Neue, Helevtica, sans-serif';
+  
+const defaultAxisLayout = () => ({
+  autorange: false,
+  showgrid: false,
+  showline: true,
+  zeroline: false,
+  mirror: true,
+  ticklen: 1,
+  tickcolor: 'white',
+  linecolor: 'white',
+  tickfont: { family: defaultFont(), size: 11, color: 'white!important' },
+  titlefont: { family: defaultFont(), size: 12, color: 'white!important' },
+  ticks: 'outside'
+});
+  
+export const defaultLayout = () => ({
+  autosize: true,
+  showgrid: false,
+  showlegend: true,
+  xaxis: defaultAxisLayout(),
+  yaxis: defaultAxisLayout(),
+  margin: { l: 50, r: 10, b: 40, t: 10 },
+  legend : {
+    xanchor : "auto",
+    yanchor : "auto",
+    font: {
+      family: defaultFont(),
+      size: 12,
+      color : '#000'
     },
-    hoverinfo: 'all',
-    type: 'scatter'
-}
-
-var defaultFont = 'Helvetica Neue, Helevtica, sans-serif';
-var defaultAxisLayout = {
-    autorange: false,
-    showgrid: false,
-    showline: true,
-    zeroline: false,
-    mirror: true,
-    ticklen: 0,
-    tickcolor: 'rgb(255, 255, 255)',
-    linecolor: 'rgb(255, 255, 255)',
-    tickfont: { family: defaultFont, size: 11, color: 'rgb(255, 255, 255)' },
-    titlefont: { family: defaultFont, size: 12, color: 'rgb(255, 255, 255)'},
-    ticks: 'outside'
-};
-var defaultLayout = {
-    autosize: true,
-    showgrid: false,
-    showlegend: true,
-    xaxis: defaultAxisLayout,
-    yaxis: defaultAxisLayout,
-    margin: { l: 50, r: 0, b: 40, t: 10 },
-    legend : {
-	xanchor : "auto",
-	yanchor : "auto",
-	font: {
-	    family: defaultFont,
-	    size: 12,
-	    color : '#fff'
-	},
-	x : 1,
-    },
-    transition: {
-	duration: 0
-    },
-    frame: {
-	duration: 0,
-	redraw: false
-    },
-    paper_bgcolor: 'rgba(50, 50, 53, 0)',
-    plot_bgcolor: 'transparent',
-    playAll : false,
-    hovermode : 'none'
-};
-
-module.exports = {
-    defaultLayout, defaultTrace
-}
+    bgcolor: '#E2E2E2',
+    bordercolor: '#AAAAAA',
+    borderwidth: 2,
+    x : 1,
+  },
+  transition: { duration: 0 },
+  frame: {
+    duration: 0,
+    redraw: false
+  },
+  paper_bgcolor: 'transparent',
+  plot_bgcolor: 'white',
+  playAll : false,
+  hovermode : 'none',
+  datarevision: 0
+});
